refactor(runtime): rename runtime/ → _runtime/
Tk patches and startup hooks are internal plumbing, not public API.
The _ prefix makes that explicit. Update 28 import references across
dialogs, style, widgets, core, and bootstack/__init__.py.
