+++ src/bootstack/_runtime/window_utilities.py
@@ -154,8 +154,8 @@ class WindowPositioning:
             to the parent.
 
         Examples:
-import bootstack.runtime.toplevel            >>> parent = tkinter.Tk()
-            >>> dialog = bootstack.runtime.toplevel.Toplevel(parent)
+import bootstack._runtime.toplevel            >>> parent = tkinter.Tk()
+            >>> dialog = bootstack._runtime.toplevel.Toplevel(parent)
             >>> dialog.update_idletasks()
             >>> parent.update_idletasks()
             >>> x, y = WindowPositioning.center_on_parent(dialog, parent)
