+++ src/bootstack/style/tk_patch.py
@@ -1,5 +1,5 @@
 ﻿"""Shim so style imports keep working while runtime owns tk_patch."""
 
-from bootstack.runtime.tk_patch import install_tk_autostyle
+from bootstack._runtime.tk_patch import install_tk_autostyle
 
 __all__ = ["install_tk_autostyle"]
