+++ src/bootstack/widgets/composites/calendar.py
@@ -14,7 +14,7 @@ from bootstack.widgets.primitives import Button, CheckToggle, Frame, Label, Sepa
 from bootstack.widgets.types import Master
 from bootstack.constants import BOTH, CENTER, LEFT, NSEW, PRIMARY, X, Y, YES
 from bootstack.core.localization import MessageCatalog
-from bootstack.runtime.utility import bind_right_click
+from bootstack._runtime.utility import bind_right_click
 from bootstack.widgets.mixins import configure_delegate
 
 ttk = SimpleNamespace(
