+++ src/bootstack/widgets/composites/tableview/tableview.py
@@ -20,7 +20,7 @@ from ttkbootstrap_icons_bs import BootstrapIcon
 from bootstack.style.style import get_style
 from bootstack.data.sqlite_source import SqliteDataSource, _ROW_ID, _ROW_SEL
 from bootstack.widgets.primitives.button import Button
-from bootstack.runtime.utility import bind_right_click
+from bootstack._runtime.utility import bind_right_click
 from bootstack.widgets.composites.contextmenu import ContextMenu
 from bootstack.widgets.composites.dropdownbutton import DropdownButton
 from bootstack.widgets.primitives.entry import Entry
